requirements.txt
agno>=1.4.0
instagrapi>=2.0.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.0
pillow>=10.3.0
rich>=13.7.0
loguru>=0.7.2
openai>=1.30.0
groq>=0.9.0
